update and delete user

+++ src/users/users.resolver.ts
@@ -1,17 +1,22 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { Resolver, Query, Mutation, Args } from '@nestjs/graphql';
 import { UsersService } from './users.service';
-import { UserInput, NewUserInput, FilterUserInput, SearchUserInput } from './user.input';
+import {
+    UserInput,
+    NewUserInput,
+    FilterUserInput,
+    SearchUserInput,
+    UpdateUserInput,
+} from './user.input';
 import { User } from './interfaces/user.entity';
 import { GqlAuthGuard, CurrentUser } from '../auth/graphql-auth.guard';
-import { UseGuards } from '@nestjs/common';
+import { UseGuards, UnauthorizedException } from '@nestjs/common';
 
-@Resolver((of) => User)
+@Resolver(() => User)
 export class UsersResolver {
     constructor(private readonly usersService: UsersService) {}
 
     @UseGuards(GqlAuthGuard)
-    @Query((returns) => User, { nullable: true })
+    @Query(() => User, { nullable: true })
     async User(
         @Args('id', { type: () => String, nullable: true }) id: string,
         @Args('email', { type: () => String, nullable: true }) email: string,
@@ -31,55 +36,52 @@ export class UsersResolver {
     }
 
     @UseGuards(GqlAuthGuard)
-    @Query((returns) => User, { nullable: true })
+    @Query(() => User, { nullable: true })
     async whoami(@CurrentUser() user: User): Promise<User> {
         return this.usersService.findOne(user._id);
     }
 
-    @Query((returns) => Boolean)
+    @Query(() => Boolean)
     async exist(@Args('email', { type: () => String }) _email: string): Promise<boolean> {
         return (await this.usersService.findOne(_email, 'email')) != null;
     }
 
     @UseGuards(GqlAuthGuard)
-    @Query((returns) => [User])
+    @Query(() => [User])
     async Users(@Args('filter', { nullable: true }) filter: FilterUserInput): Promise<User[]> {
         return this.usersService.findAll(filter);
     }
 
     @UseGuards(GqlAuthGuard)
-    @Mutation((returns) => User)
+    @Mutation(() => User)
     async createUser(@Args('input') input: UserInput): Promise<User> {
         return this.usersService.create(input as User);
     }
 
-    @Mutation((returns) => User)
+    @Mutation(() => User)
     async newUser(@Args('input') input: NewUserInput): Promise<User> {
         return this.usersService.newUser(input);
     }
-    /*
-    @UseGuards(GqlAuthGuard)
-    @Mutation((returns) => UserType, { nullable: true })
-    async updateUser(
-        @Args('id') id: string,
-        @Args('input') input: UpdateUserInput
-    ): Promise<UserType> {
-        return this.usersService.update(id, input as User);
-    }
 
     @UseGuards(GqlAuthGuard)
-    @Mutation((returns) => UserType, { nullable: true })
-    async deleteUser(@Args('id') id: string): Promise<UserType> {
+    @Mutation(() => User, { nullable: true })
+    async deleteUser(@CurrentUser() user: User, @Args('id') id: string): Promise<User> {
+        if (user._id != id && user.credidential < 8) {
+            throw new UnauthorizedException('You are not allowed to delete this user');
+        }
         return this.usersService.delete(id);
     }
 
     @UseGuards(GqlAuthGuard)
-    @Mutation((returns) => [UserType])
-    async deleteByCriteria(@Args('filter') filter: FilterUserInput): Promise<UserType[]> {
-        const result = await this.usersService.findByCriteria(filter);
-        result.forEach((element) => {
-            this.usersService.delete(element.id);
-        });
-        return result;
-    }*/
+    @Mutation(() => User, { nullable: true })
+    async updateUser(
+        @CurrentUser() user: User,
+        @Args('id') id: string,
+        @Args('input') input: UpdateUserInput
+    ): Promise<User> {
+        if (user._id != id && user.credidential < 8) {
+            throw new UnauthorizedException('You are not allowed to update this user');
+        }
+        return this.usersService.update(id, input as User);
+    }
 }
